adding editing changes
