+++ src/main/java/db/connectionpool/ConnectionPool.java
@@ -1,4 +1,4 @@
-package db.connectionpoolm;
+package db.connectionpool;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
