+++ src/main/java/db/dao/IHotelDAO.java
