+++ src/main/java/db/dao/mysql/FlightDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.IPaymentTypeDAO;
-import db.models.PaymentType;
+import db.connectionpool.ConnectionPool;
+import db.dao.IFlightDAO;
+import db.models.Flight;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,21 +12,24 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class PaymentTypeDAO implements IPaymentTypeDAO {
-    private static final Logger LOGGER = LogManager.getLogger(PaymentTypeDAO.class);
+public class FlightDAO implements IFlightDAO {
+    private static final Logger LOGGER = LogManager.getLogger(FlightDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<PaymentType> getAllEntities(){
-        List<PaymentType>  ptList = new ArrayList<>();
-        String sql = "SELECT * FROM payment_types";
+    public List<Flight> getAllEntities(){
+        List<Flight>  flightList = new ArrayList<>();
+        String sql = "SELECT * FROM flights";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                PaymentType pt = new PaymentType();
-                pt.setPaymentType(rs.getInt(1));
-                pt.setPaymentTypeName(rs.getString(2));
-                ptList.add(pt);
+                Flight flight = new Flight();
+                flight.setFlightId(rs.getInt(1));
+                flight.setPrice(rs.getDouble(2));
+                flight.setAirlineId(rs.getInt(3));
+                flight.setOriginCityId(rs.getInt(4));
+                flight.setDestinationCityId(rs.getInt(5));
+                flightList.add(flight);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -39,20 +42,23 @@ public class PaymentTypeDAO implements IPaymentTypeDAO {
                 }
             }
         }
-        return ptList;
+        return flightList;
     }
 
-    public PaymentType getEntityById(int id) {
-        PaymentType pt = null;
+    public Flight getEntityById(int id) {
+        Flight flight = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM payment_type WHERE payment_type_id = (?)";
+        String sql = "SELECT * FROM flights WHERE flight_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int paymentTypeId = rs.getInt("payment_type_id");
-                String ptName = rs.getString("payment_type_name");
-                pt = new PaymentType(paymentTypeId, ptName);
+                int flightId = rs.getInt("flight_id");
+                double price = rs.getDouble("price");
+                int airlineId = rs.getInt("airline_id");
+                int originCityId = rs.getInt("origin_city_id");
+                int destinationCityId = rs.getInt("destination_city_id");
+                flight = new Flight(flightId, price, airlineId, originCityId, destinationCityId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -65,20 +71,23 @@ public class PaymentTypeDAO implements IPaymentTypeDAO {
                 }
             }
         }
-        return pt;
+        return flight;
     }
 
-    public PaymentType getPaymentTypeByName(String dbName) {
-        PaymentType pt = null;
+    public Flight getFlightPrice(double dbPrice) {
+        Flight flight = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM payment_type WHERE payment_type_name = (?)";
+        String sql = "SELECT * FROM flights WHERE price = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setString(1, dbName);
+            ps.setDouble(1, dbPrice);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int paymentTypeId = rs.getInt("payment_type_id");
-                String ptName = rs.getString("payment_type_name");
-                pt = new PaymentType(paymentTypeId, ptName);
+                int flightId = rs.getInt("flight_id");
+                double price = rs.getDouble("price");
+                int airlineId = rs.getInt("airline_id");
+                int originCityId = rs.getInt("origin_city_id");
+                int destinationCityId = rs.getInt("destination_city_id");
+                flight = new Flight(flightId, price, airlineId, originCityId, destinationCityId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -91,16 +100,18 @@ public class PaymentTypeDAO implements IPaymentTypeDAO {
                 }
             }
         }
-        return pt;
+        return flight;
     }
 
-    public PaymentType createEntity(PaymentType pt) {
+    public Flight createEntity(Flight flight) {
         Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO payment_types (payment_type_id, payment_type_name) " +
-                "VALUES (?,?)";
+        String sql = "INSERT INTO flights (flight_id, price, airline_id, origin_city_id, destination_city_id) VALUES (?,?,?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, pt.getPaymentType());
-            ps.setString(2, pt.getPaymentTypeName());
+            ps.setInt(1, flight.getFlightId());
+            ps.setDouble(2, flight.getPrice());
+            ps.setInt(3, flight.getAirlineId());
+            ps.setInt(4, flight.getOriginCityId());
+            ps.setInt(5, flight.getDestinationCityId());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -117,13 +128,17 @@ public class PaymentTypeDAO implements IPaymentTypeDAO {
         return null;
     }
 
-    public void updateEntity(PaymentType pt) {
-        String sql = "UPDATE payment_types SET payment_type_name = (?) WHERE payment_type_id = (?)";
+    public void updateEntity(Flight flight) {
+        String sql = "UPDATE flights SET price = (?), airline_id = (?), origin_city_id = (?), " +
+                     "destination_city_id = (?) WHERE flight_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + pt);
-            ps.setString(1, pt.getPaymentTypeName());
-            ps.setInt(2, pt.getPaymentType());
+            LOGGER.info("in update customer: " + flight);
+            ps.setDouble(1, flight.getPrice());
+            ps.setInt(2, flight.getAirlineId());
+            ps.setInt(3, flight.getOriginCityId());
+            ps.setInt(4, flight.getDestinationCityId());
+            ps.setInt(5, flight.getFlightId());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
@@ -139,7 +154,7 @@ public class PaymentTypeDAO implements IPaymentTypeDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM payment_types WHERE payment_type_id = (?)";
+        String sql = "Delete FROM flights WHERE flight_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
