+++ src/main/java/db/dao/mysql/HotelDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.IFlightDAO;
-import db.models.Flight;
+import db.connectionpool.ConnectionPool;
+import db.dao.IHotelDAO;
+import db.models.Hotel;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,24 +12,23 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FlightDAO implements IFlightDAO {
-    private static final Logger LOGGER = LogManager.getLogger(FlightDAO.class);
+public class HotelDAO implements IHotelDAO {
+    private static final Logger LOGGER = LogManager.getLogger(HotelDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<Flight> getAllEntities(){
-        List<Flight>  flightList = new ArrayList<>();
-        String sql = "SELECT * FROM flights";
+    public List<Hotel> getAllEntities(){
+        List<Hotel>  hotelList = new ArrayList<>();
+        String sql = "SELECT * FROM hotels";
         Connection con = connectionPool.getConnection();
-        try (PreparedStatement ps = con.prepareStatement(sql)) {
+        try (PreparedStatement ps = con.prepareStatement(sql)){
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Flight flight = new Flight();
-                flight.setFlightId(rs.getInt(1));
-                flight.setPrice(rs.getDouble(2));
-                flight.setAirlineId(rs.getInt(3));
-                flight.setOriginCityId(rs.getInt(4));
-                flight.setDestinationCityId(rs.getInt(5));
-                flightList.add(flight);
+                Hotel hotel = new Hotel();
+                hotel.setHotelId(rs.getInt(1));
+                hotel.setHotelName(rs.getString(2));
+                hotel.setPrice(rs.getDouble(3));
+                hotel.setCityId(rs.getInt(4));
+                hotelList.add(hotel);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -42,23 +41,22 @@ public class FlightDAO implements IFlightDAO {
                 }
             }
         }
-        return flightList;
+        return hotelList;
     }
 
-    public Flight getEntityById(int id) {
-        Flight flight = null;
+    public Hotel getEntityById(int id) {
+        Hotel hotel = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM flights WHERE flight_id = (?)";
+        String sql = "SELECT * FROM hotels WHERE hotel_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int flightId = rs.getInt("flight_id");
+                int hotelId = rs.getInt("hotel_id");
+                String hotelName = rs.getString("hotel_name");
                 double price = rs.getDouble("price");
-                int airlineId = rs.getInt("airline_id");
-                int originCityId = rs.getInt("origin_city_id");
-                int destinationCityId = rs.getInt("destination_city_id");
-                flight = new Flight(flightId, price, airlineId, originCityId, destinationCityId);
+                int cityId = rs.getInt("city_id");
+                hotel = new Hotel(hotelId, hotelName, price, cityId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -71,23 +69,50 @@ public class FlightDAO implements IFlightDAO {
                 }
             }
         }
-        return flight;
+        return hotel;
     }
 
-    public Flight getFlightPrice(double dbPrice) {
-        Flight flight = null;
+    public Hotel getHotelByName(String dbName) {
+        Hotel hotel = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM flights WHERE price = (?)";
+        String sql = "SELECT * FROM hotels WHERE hotel_name = (?)";
+        try (PreparedStatement ps = con.prepareStatement(sql)) {
+            ps.setString(1, dbName);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                int hotelId = rs.getInt("hotel_id");
+                String hotelName = rs.getString("hotel_name");
+                double price = rs.getDouble("price");
+                int cityId = rs.getInt("city_id");
+                hotel = new Hotel(hotelId, hotelName, price, cityId);
+            }
+        } catch (Exception e) {
+            LOGGER.error(e);
+        } finally {
+            if (con != null) {
+                try {
+                    connectionPool.releaseConnection(con);
+                } catch (Exception e) {
+                    LOGGER.info(e);
+                }
+            }
+        }
+        return hotel;
+    }
+
+    public Hotel getHotelByPrice(double dbPrice) {
+        Hotel hotel = null;
+        Connection con = connectionPool.getConnection();
+        String sql = "SELECT * FROM hotels WHERE price = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setDouble(1, dbPrice);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int flightId = rs.getInt("flight_id");
+                int hotelId = rs.getInt("hotel_id");
+                String hotelName = rs.getString("hotel_name");
                 double price = rs.getDouble("price");
-                int airlineId = rs.getInt("airline_id");
-                int originCityId = rs.getInt("origin_city_id");
-                int destinationCityId = rs.getInt("destination_city_id");
-                flight = new Flight(flightId, price, airlineId, originCityId, destinationCityId);
+                int cityId = rs.getInt("city_id");
+                hotel = new Hotel(hotelId, hotelName, price, cityId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -100,18 +125,17 @@ public class FlightDAO implements IFlightDAO {
                 }
             }
         }
-        return flight;
+        return hotel;
     }
 
-    public Flight createEntity(Flight flight) {
+    public Hotel createEntity(Hotel hotel) {
         Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO flights (flight_id, price, airline_id, origin_city_id, destination_city_id) VALUES (?,?,?,?,?)";
+        String sql = "INSERT INTO hotels (hotel_id, hotel_name, price, city_id) VALUES (?,?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, flight.getFlightId());
-            ps.setDouble(2, flight.getPrice());
-            ps.setInt(3, flight.getAirlineId());
-            ps.setInt(4, flight.getOriginCityId());
-            ps.setInt(5, flight.getDestinationCityId());
+            ps.setInt(1, hotel.getHotelId());
+            ps.setString(2, hotel.getHotelName());
+            ps.setDouble(3, hotel.getPrice());
+            ps.setInt(4, hotel.getCityId());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -128,17 +152,16 @@ public class FlightDAO implements IFlightDAO {
         return null;
     }
 
-    public void updateEntity(Flight flight) {
-        String sql = "UPDATE flights SET price = (?), airline_id = (?), origin_city_id = (?), " +
-                     "destination_city_id = (?) WHERE flight_id = (?)";
+    public void updateEntity(Hotel hotel) {
+        String sql = "UPDATE hotels SET hotel_name = (?), price = (?), city_id = (?), " +
+                "WHERE hotel_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + flight);
-            ps.setDouble(1, flight.getPrice());
-            ps.setInt(2, flight.getAirlineId());
-            ps.setInt(3, flight.getOriginCityId());
-            ps.setInt(4, flight.getDestinationCityId());
-            ps.setInt(5, flight.getFlightId());
+            LOGGER.info("in update customer: " + hotel);
+            ps.setString(1, hotel.getHotelName());
+            ps.setDouble(2, hotel.getPrice());
+            ps.setInt(3, hotel.getCityId());
+            ps.setInt(4, hotel.getCityId());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
@@ -154,7 +177,7 @@ public class FlightDAO implements IFlightDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM flights WHERE flight_id = (?)";
+        String sql = "Delete FROM hotels WHERE hotel_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
