+++ src/main/java/db/dao/mysql/OrderDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.IDrivingCompanyDAO;
-import db.models.DrivingCompany;
+import db.connectionpool.ConnectionPool;
+import db.dao.IOrderDAO;
+import db.models.Order;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,22 +12,27 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DrivingCompanyDAO implements IDrivingCompanyDAO {
-    private static final Logger LOGGER = LogManager.getLogger(DrivingCompanyDAO.class);
+public class OrderDAO implements IOrderDAO {
+    private static final Logger LOGGER = LogManager.getLogger(OrderDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<DrivingCompany> getAllEntities(){
-        List<DrivingCompany>  dcList = new ArrayList<>();
-        String sql = "SELECT * FROM driving_companies";
+    public List<Order> getAllEntities(){
+        List<Order>  orderList = new ArrayList<>();
+        String sql = "SELECT * FROM orders";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                DrivingCompany dc = new DrivingCompany();
-                dc.setDrivingId(rs.getInt(1));
-                dc.setPrice(rs.getDouble(2));
-                dc.setCityId(rs.getInt(3));
-                dcList.add(dc);
+                Order order = new Order();
+                order.setOrderId(rs.getInt(1));
+                order.setPayment(rs.getDouble(2));
+                order.setCustomerId(rs.getInt(3));
+                order.setStaffId(rs.getInt(4));
+                order.setToursId(rs.getInt(5));
+                order.setStatusId(rs.getInt(6));
+                order.setPaymentId(rs.getInt(7));
+                order.setDrivingCompaniesId(rs.getInt(8));
+                orderList.add(order);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -40,21 +45,26 @@ public class DrivingCompanyDAO implements IDrivingCompanyDAO {
                 }
             }
         }
-        return dcList;
+        return orderList;
     }
 
-    public DrivingCompany getEntityById(int id) {
-        DrivingCompany dc = null;
+    public Order getEntityById(int id) {
+        Order order = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM driving_companies WHERE driving_id = (?)";
+        String sql = "SELECT * FROM orders WHERE order_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
+                int orderId = rs.getInt("order_id");
+                double payment = rs.getDouble("payment");
+                int customerId = rs.getInt("customer_id");
+                int staffId = rs.getInt("staff_id");
+                int toursId = rs.getInt("tours_id");
+                int statusId = rs.getInt("status_id");
+                int paymentId = rs.getInt("payment_id");
                 int drivingId = rs.getInt("driving_id");
-                double price = rs.getDouble("price");
-                int cityId = rs.getInt("city_id");
-                dc = new DrivingCompany(drivingId, price, cityId);
+                order = new Order(orderId, payment, customerId, staffId, toursId, statusId, paymentId, drivingId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -67,21 +77,26 @@ public class DrivingCompanyDAO implements IDrivingCompanyDAO {
                 }
             }
         }
-        return dc;
+        return order;
     }
 
-    public DrivingCompany getDrivingCompanyByPrice(double dbPrice) {
-        DrivingCompany dc = null;
+    public Order getOrderByPayment(int dbPayment) {
+        Order order = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM driving_companies WHERE price = (?)";
+        String sql = "SELECT * FROM orders WHERE payment = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setDouble(1, dbPrice);
+            ps.setDouble(1, dbPayment);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
+                int orderId = rs.getInt("order_id");
+                double payment = rs.getDouble("payment");
+                int customerId = rs.getInt("customer_id");
+                int staffId = rs.getInt("staff_id");
+                int toursId = rs.getInt("tours_id");
+                int statusId = rs.getInt("status_id");
+                int paymentId = rs.getInt("payment_id");
                 int drivingId = rs.getInt("driving_id");
-                double price = rs.getDouble("price");
-                int cityId = rs.getInt("city_id");
-                dc = new DrivingCompany(drivingId, price, cityId);
+                order = new Order(orderId, payment, customerId, staffId, toursId, statusId, paymentId, drivingId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -94,16 +109,22 @@ public class DrivingCompanyDAO implements IDrivingCompanyDAO {
                 }
             }
         }
-        return dc;
+        return order;
     }
 
-    public DrivingCompany createEntity(DrivingCompany dc) {
+    public Order createEntity(Order order) {
         Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO driving_companies (driving_id, price, city_id) VALUES (?,?,?)";
+        String sql = "INSERT INTO hotels (order_id, payment, customer_id, staff_id, tours_id, status_id, payment_id, driving_Id) " +
+                     "VALUES (?,?,?,?,?,?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, dc.getDrivingId());
-            ps.setDouble(2, dc.getPrice());
-            ps.setInt(3, dc.getCityId());
+            ps.setInt(1, order.getOrderId());
+            ps.setDouble(2, order.getPayment());
+            ps.setInt(3, order.getCustomerId());
+            ps.setInt(4, order.getStaffId());
+            ps.setInt(5, order.getToursId());
+            ps.setInt(6, order.getStatusId());
+            ps.setInt(7, order.getPaymentId());
+            ps.setInt(8, order.getDrivingCompaniesId());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -120,14 +141,20 @@ public class DrivingCompanyDAO implements IDrivingCompanyDAO {
         return null;
     }
 
-    public void updateEntity(DrivingCompany dc) {
-        String sql = "UPDATE driving_companies SET price = (?), city_id = (?) WHERE driving_id = (?)";
+    public void updateEntity(Order order) {
+        String sql = "UPDATE orders SET payment = (?), customer_id = (?), staff_id = (?), " +
+                "tours_id = ?, status_id = ?, payment_id = ?, driving_id = ? WHERE order_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + dc);
-            ps.setDouble(1, dc.getPrice());
-            ps.setInt(2, dc.getCityId());
-            ps.setInt(3, dc.getDrivingId());
+            LOGGER.info("in update customer: " + order);
+            ps.setDouble(1, order.getPayment());
+            ps.setInt(2, order.getCustomerId());
+            ps.setInt(3, order.getStaffId());
+            ps.setInt(4, order.getToursId());
+            ps.setInt(5, order.getStatusId());
+            ps.setInt(6, order.getPaymentId());
+            ps.setInt(7, order.getDrivingCompaniesId());
+            ps.setInt(8, order.getOrderId());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
@@ -143,7 +170,7 @@ public class DrivingCompanyDAO implements IDrivingCompanyDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM driving_companies WHERE driving_id = (?)";
+        String sql = "Delete FROM orders WHERE order_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
