+++ src/main/java/db/dao/mysql/PaymentTypeDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.ITourDAO;
-import db.models.Tour;
+import db.connectionpool.ConnectionPool;
+import db.dao.IPaymentTypeDAO;
+import db.models.PaymentType;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,24 +12,21 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class TourDAO implements ITourDAO {
-    private static final Logger LOGGER = LogManager.getLogger(TourDAO.class);
+public class PaymentTypeDAO implements IPaymentTypeDAO {
+    private static final Logger LOGGER = LogManager.getLogger(PaymentTypeDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<Tour> getAllEntities(){
-        List<Tour>  tourList = new ArrayList<>();
-        String sql = "SELECT * FROM tours";
+    public List<PaymentType> getAllEntities(){
+        List<PaymentType>  ptList = new ArrayList<>();
+        String sql = "SELECT * FROM payment_types";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Tour tour = new Tour();
-                tour.setTourId(rs.getInt(1));
-                tour.setTourName(rs.getString(2));
-                tour.setHotelId(rs.getInt(3));
-                tour.setFlightToId(rs.getInt(4));
-                tour.setFlightFromId(rs.getInt(5));
-                tourList.add(tour);
+                PaymentType pt = new PaymentType();
+                pt.setPaymentType(rs.getInt(1));
+                pt.setPaymentTypeName(rs.getString(2));
+                ptList.add(pt);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -42,23 +39,20 @@ public class TourDAO implements ITourDAO {
                 }
             }
         }
-        return tourList;
+        return ptList;
     }
 
-    public Tour getEntityById(int id) {
-        Tour tour = null;
+    public PaymentType getEntityById(int id) {
+        PaymentType pt = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM tours WHERE tour_id = (?)";
+        String sql = "SELECT * FROM payment_type WHERE payment_type_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int tourId = rs.getInt("tour_id");
-                String tourName = rs.getString("tour_name");
-                int hotelId = rs.getInt("hotel_id");
-                int flightToId = rs.getInt("flight_to_id");
-                int flightFromId = rs.getInt("flight_from_id");
-                tour = new Tour(tourId, tourName, hotelId,flightToId, flightFromId);
+                int paymentTypeId = rs.getInt("payment_type_id");
+                String ptName = rs.getString("payment_type_name");
+                pt = new PaymentType(paymentTypeId, ptName);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -71,23 +65,20 @@ public class TourDAO implements ITourDAO {
                 }
             }
         }
-        return tour;
+        return pt;
     }
 
-    public Tour getTourByName(String dbName) {
-        Tour tour = null;
+    public PaymentType getPaymentTypeByName(String dbName) {
+        PaymentType pt = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM tours WHERE tour_name = (?)";
+        String sql = "SELECT * FROM payment_type WHERE payment_type_name = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setString(1, dbName);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int tourId = rs.getInt("tour_id");
-                String tourName = rs.getString("tour_name");
-                int hotelId = rs.getInt("hotel_id");
-                int flightToId = rs.getInt("flight_to_id");
-                int flightFromId = rs.getInt("flight_from_id");
-                tour = new Tour(tourId, tourName, hotelId,flightToId, flightFromId);
+                int paymentTypeId = rs.getInt("payment_type_id");
+                String ptName = rs.getString("payment_type_name");
+                pt = new PaymentType(paymentTypeId, ptName);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -100,19 +91,16 @@ public class TourDAO implements ITourDAO {
                 }
             }
         }
-        return tour;
+        return pt;
     }
 
-    public Tour createEntity(Tour tour) {
+    public PaymentType createEntity(PaymentType pt) {
         Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO tours (tour_id, tour_name, hotel_id, flight_to_id, flight_from_id) " +
-                "VALUES (?,?,?,?,?)";
+        String sql = "INSERT INTO payment_types (payment_type_id, payment_type_name) " +
+                "VALUES (?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, tour.getTourId());
-            ps.setString(2, tour.getTourName());
-            ps.setInt(3, tour.getHotelId());
-            ps.setInt(4, tour.getFlightToId());
-            ps.setInt(5, tour.getFlightFromId());
+            ps.setInt(1, pt.getPaymentType());
+            ps.setString(2, pt.getPaymentTypeName());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -129,17 +117,13 @@ public class TourDAO implements ITourDAO {
         return null;
     }
 
-    public void updateEntity(Tour tour) {
-        String sql = "UPDATE tours SET tour_name = (?), hotel_id = (?)" +
-                "flight_to_id = (?), flight_from_id = (?) WHERE tour_id = (?)";
+    public void updateEntity(PaymentType pt) {
+        String sql = "UPDATE payment_types SET payment_type_name = (?) WHERE payment_type_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + tour);
-            ps.setString(1, tour.getTourName());
-            ps.setInt(2, tour.getHotelId());
-            ps.setInt(3, tour.getFlightToId());
-            ps.setInt(4, tour.getFlightFromId());
-            ps.setInt(5, tour.getTourId());
+            LOGGER.info("in update customer: " + pt);
+            ps.setString(1, pt.getPaymentTypeName());
+            ps.setInt(2, pt.getPaymentType());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
@@ -155,7 +139,7 @@ public class TourDAO implements ITourDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM tours WHERE tour_id = (?)";
+        String sql = "Delete FROM payment_types WHERE payment_type_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
