+++ src/main/java/db/dao/mysql/StaffDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.ICustomerDAO;
-import db.models.Customer;
+import db.connectionpool.ConnectionPool;
+import db.dao.IStaffDAO;
+import db.models.Staff;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,24 +12,22 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CustomerDAO implements ICustomerDAO {
-
-    private static final Logger LOGGER = LogManager.getLogger(CustomerDAO.class);
+public class StaffDAO implements IStaffDAO {
+    private static final Logger LOGGER = LogManager.getLogger(StaffDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<Customer> getAllEntities(){
-        List<Customer> customerList = new ArrayList<>();
-        String sql = "SELECT customer_id, first_name, last_name, phone FROM customers";
+    public List<Staff> getAllEntities(){
+        List<Staff> staffList = new ArrayList<Staff>();
+        String sql = "SELECT * FROM staff";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Customer customer = new Customer();
-                customer.setCustomerId(rs.getInt(1));
-                customer.setFirstName(rs.getString(2));
-                customer.setLastName(rs.getString(3));
-                customer.setPhone(rs.getString(4));
-                customerList.add(customer);
+                Staff staff = new Staff();
+                staff.setStaffId(rs.getInt(1));
+                staff.setFirstName(rs.getString(2));
+                staff.setLastName(rs.getString(3));
+                staffList.add(staff);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -42,23 +40,21 @@ public class CustomerDAO implements ICustomerDAO {
                 }
             }
         }
-
-        return customerList;
+        return staffList;
     }
 
-   public Customer getEntityById(int id) {
-        Customer customers = null;
+    public Staff getEntityById(int id) {
+        Staff staff = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM customers WHERE customer_id = (?)";
+        String sql = "SELECT * FROM staff WHERE staff_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int customerId = rs.getInt("customer_id");
+                int customerId = rs.getInt("staff_id");
                 String firstName = rs.getString("first_name");
                 String lastName = rs.getString("last_name");
-                String phone = rs.getString("phone");
-                customers = new Customer(customerId, firstName, lastName, phone);
+                staff = new Staff(customerId, firstName, lastName);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -71,22 +67,48 @@ public class CustomerDAO implements ICustomerDAO {
                 }
             }
         }
-        return customers;
+        return staff;
     }
 
-    public Customer getCustomerByFirstName(String dbLastName) {
-        Customer customers = null;
+    public Staff getStaffByFirstName(String dbFirstName) {
+        Staff staff = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM customers WHERE first_name = (?)";
+        String sql = "SELECT * FROM staff WHERE first_name = (?)";
+        try (PreparedStatement ps = con.prepareStatement(sql)) {
+            ps.setString(1, dbFirstName);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                int customerId = rs.getInt("staff_id");
+                String firstName = rs.getString("first_name");
+                String lastName = rs.getString("last_name");
+                staff = new Staff(customerId, firstName, lastName);
+            }
+        } catch (Exception e) {
+            LOGGER.error(e);
+        } finally {
+            if (con != null) {
+                try {
+                    connectionPool.releaseConnection(con);
+                } catch (Exception e) {
+                    LOGGER.info(e);
+                }
+            }
+        }
+        return staff;
+    }
+
+    public Staff getStaffByLastName(String dbLastName) {
+        Staff staff = null;
+        Connection con = connectionPool.getConnection();
+        String sql = "SELECT * FROM staff WHERE last_name = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setString(1, dbLastName);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int customerId = rs.getInt("customer_id");
+                int customerId = rs.getInt("staff_id");
                 String firstName = rs.getString("first_name");
                 String lastName = rs.getString("last_name");
-                String phone = rs.getString("phone");
-                customers = new Customer(customerId, firstName, lastName, phone);
+                staff = new Staff(customerId, firstName, lastName);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -99,73 +121,16 @@ public class CustomerDAO implements ICustomerDAO {
                 }
             }
         }
-        return customers;
+        return staff;
     }
 
-    public Customer getCustomerByLastName(String dbLastName) {
-        Customer customers = null;
+    public Staff createEntity(Staff staff) {
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM customers WHERE last_name = (?)";
+        String sql = "INSERT INTO staff (staff_id, first_name, last_name) VALUES (?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setString(1, dbLastName);
-            ResultSet rs = ps.executeQuery();
-            if (rs.next()) {
-                int customerId = rs.getInt("customer_id");
-                String firstName = rs.getString("first_name");
-                String lastName = rs.getString("last_name");
-                String phone = rs.getString("phone");
-                customers = new Customer(customerId, firstName, lastName, phone);
-            }
-        } catch (Exception e) {
-            LOGGER.error(e);
-        } finally {
-            if (con != null) {
-                try {
-                    connectionPool.releaseConnection(con);
-                } catch (Exception e) {
-                    LOGGER.info(e);
-                }
-            }
-        }
-        return customers;
-    }
-
-    public Customer getCustomerByPhone(String dbPhone) {
-        Customer customers = null;
-        Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM customers WHERE phone = (?)";
-        try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setString(1, dbPhone);
-            ResultSet rs = ps.executeQuery();
-            if (rs.next()) {
-                int customerId = rs.getInt("customer_id");
-                String firstName = rs.getString("first_name");
-                String lastName = rs.getString("last_name");
-                String phone = rs.getString("phone");
-                customers = new Customer(customerId, firstName, lastName, phone);
-            }
-        } catch (Exception e) {
-            LOGGER.error(e);
-        } finally {
-            if (con != null) {
-                try {
-                    connectionPool.releaseConnection(con);
-                } catch (Exception e) {
-                    LOGGER.info(e);
-                }
-            }
-        }
-        return customers;
-    }
-
-    public Customer createEntity(Customer customers) {
-        Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO customers (customer_id, first_name, last_name, phone) VALUES (?,?,?,?)";
-        try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, customers.getCustomerId());
-            ps.setString(2, customers.getFirstName());
-            ps.setString(3, customers.getLastName());
-            ps.setString(4, customers.getPhone());
+            ps.setInt(1, staff.getStaffId());
+            ps.setString(2, staff.getFirstName());
+            ps.setString(3, staff.getLastName());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -182,16 +147,15 @@ public class CustomerDAO implements ICustomerDAO {
         return null;
     }
 
-    public void updateEntity(Customer customers) {
-        String sql = "UPDATE customers SET first_name = (?), last_name = (?), phone = (?) WHERE customer_id = (?)";
+    public void updateEntity(Staff staff) {
+        String sql = "UPDATE staff SET first_name = ?, last_name = ? WHERE customer_id = ?";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + customers);
-            ps.setString(1, customers.getFirstName());
-            ps.setString(2, customers.getLastName());
-            ps.setString(3, customers.getPhone());
-            ps.setInt(4, customers.getCustomerId());
-            ps.execute();
+            LOGGER.info("in update staff: " + staff);
+            ps.setString(1, staff.getFirstName());
+            ps.setString(2, staff.getLastName());
+            ps.setInt(3, staff.getStaffId());
+            ps.executeUpdate();
         } catch(Exception e) {
             LOGGER.error(e);
         } finally {
@@ -206,7 +170,7 @@ public class CustomerDAO implements ICustomerDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM customers WHERE customer_id = ?";
+        String sql = "Delete FROM staff WHERE staff_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
