+++ src/main/java/db/dao/mysql/TourDAO.java
@@ -1,8 +1,8 @@
 package db.dao.mysql;
 
-import db.connectionpoolm.ConnectionPool;
-import db.dao.ICountryDAO;
-import db.models.Country;
+import db.connectionpool.ConnectionPool;
+import db.dao.ITourDAO;
+import db.models.Tour;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,21 +12,24 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CountryDAO implements ICountryDAO {
-    private static final Logger LOGGER = LogManager.getLogger(CountryDAO.class);
+public class TourDAO implements ITourDAO {
+    private static final Logger LOGGER = LogManager.getLogger(TourDAO.class);
     private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<Country> getAllEntities(){
-        List<Country> countryList = new ArrayList<>();
-        String sql = "SELECT * FROM countries";
-        Connection con = connectionPool.getConnection();;
+    public List<Tour> getAllEntities(){
+        List<Tour>  tourList = new ArrayList<>();
+        String sql = "SELECT * FROM tours";
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Country country = new Country();
-                country.setCountryId(rs.getInt(1));
-                country.setCountryName(rs.getString(2));
-                countryList.add(country);
+                Tour tour = new Tour();
+                tour.setTourId(rs.getInt(1));
+                tour.setTourName(rs.getString(2));
+                tour.setHotelId(rs.getInt(3));
+                tour.setFlightToId(rs.getInt(4));
+                tour.setFlightFromId(rs.getInt(5));
+                tourList.add(tour);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -39,20 +42,23 @@ public class CountryDAO implements ICountryDAO {
                 }
             }
         }
-        return countryList;
+        return tourList;
     }
 
-    public Country getEntityById(int id) {
-        Country country = null;
+    public Tour getEntityById(int id) {
+        Tour tour = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM countries WHERE country_id = (?)";
+        String sql = "SELECT * FROM tours WHERE tour_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int countryId = rs.getInt("country_id");
-                String countryName = rs.getString("country_name");
-                country = new Country(countryId, countryName);
+                int tourId = rs.getInt("tour_id");
+                String tourName = rs.getString("tour_name");
+                int hotelId = rs.getInt("hotel_id");
+                int flightToId = rs.getInt("flight_to_id");
+                int flightFromId = rs.getInt("flight_from_id");
+                tour = new Tour(tourId, tourName, hotelId,flightToId, flightFromId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -65,20 +71,23 @@ public class CountryDAO implements ICountryDAO {
                 }
             }
         }
-        return country;
+        return tour;
     }
 
-    public Country getCountryByName(String dbName) {
-        Country country = null;
+    public Tour getTourByName(String dbName) {
+        Tour tour = null;
         Connection con = connectionPool.getConnection();
-        String sql = "SELECT * FROM countries WHERE country_name = (?)";
+        String sql = "SELECT * FROM tours WHERE tour_name = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setString(1, dbName);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int countryId = rs.getInt("country_id");
-                String countryName = rs.getString("country_name");
-                country = new Country(countryId, countryName);
+                int tourId = rs.getInt("tour_id");
+                String tourName = rs.getString("tour_name");
+                int hotelId = rs.getInt("hotel_id");
+                int flightToId = rs.getInt("flight_to_id");
+                int flightFromId = rs.getInt("flight_from_id");
+                tour = new Tour(tourId, tourName, hotelId,flightToId, flightFromId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
@@ -91,15 +100,19 @@ public class CountryDAO implements ICountryDAO {
                 }
             }
         }
-        return country;
+        return tour;
     }
 
-    public Country createEntity(Country country) {
+    public Tour createEntity(Tour tour) {
         Connection con = connectionPool.getConnection();
-        String sql = "INSERT INTO countries (country_id, country_name) VALUES (?,?)";
+        String sql = "INSERT INTO tours (tour_id, tour_name, hotel_id, flight_to_id, flight_from_id) " +
+                "VALUES (?,?,?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            ps.setInt(1, country.getCountryId());
-            ps.setString(2, country.getCountryName());
+            ps.setInt(1, tour.getTourId());
+            ps.setString(2, tour.getTourName());
+            ps.setInt(3, tour.getHotelId());
+            ps.setInt(4, tour.getFlightToId());
+            ps.setInt(5, tour.getFlightFromId());
             ps.executeUpdate();
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
@@ -116,13 +129,17 @@ public class CountryDAO implements ICountryDAO {
         return null;
     }
 
-    public void updateEntity(Country country) {
-        String sql = "UPDATE countries SET country_name = (?) WHERE country_id = (?)";
+    public void updateEntity(Tour tour) {
+        String sql = "UPDATE tours SET tour_name = (?), hotel_id = (?)" +
+                "flight_to_id = (?), flight_from_id = (?) WHERE tour_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-            LOGGER.info("in update customer: " + country);
-            ps.setString(1, country.getCountryName());
-            ps.setInt(2, country.getCountryId());
+            LOGGER.info("in update customer: " + tour);
+            ps.setString(1, tour.getTourName());
+            ps.setInt(2, tour.getHotelId());
+            ps.setInt(3, tour.getFlightToId());
+            ps.setInt(4, tour.getFlightFromId());
+            ps.setInt(5, tour.getTourId());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
@@ -138,7 +155,7 @@ public class CountryDAO implements ICountryDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM countries WHERE country_id = (?)";
+        String sql = "Delete FROM tours WHERE tour_id = (?)";
         Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
